ticks for tracing
We reuse HPC ticks for generating stack traces.
For this we add a boolean flag to the HsTick constructor that denotes
whether, at runtime, this tick should update the current trace center.
In compiled mode and since the coverage way is not compatible with the
trace way, a tick which does not have this flag set is a noop (and we
should avoid inserting it in the first place)
In interpreted mode the story is different and a tick may give rise both
to a breakpoint and a trace center update. Bear in mind that we do not want
to update the current trace center for *every* breakpoint, as that would be
wasteful (or maybe not?).