Initialize CCSs from native shared object
The CCSs are registered to a pending list CCS_LIST by the library constructor
functions, which is only properly initilized by calling initProfiling2 later.
We need do so in loadNativeObj as well, otherwise, the CCSs will be in
invlid state (root == nullptr) and profiling code path can crash.

**NOTE [Notice to future ghc maintainer]**

Please squash this commit with:

https://github.com/facebook/fbghc/commit/74d8cadeef8a509f1ee612795f5af6142edb006c
https://github.com/facebook/fbghc/commit/77002f1b2041e2bafd6b1e45c015f53e19cf6371